Refactoring Website Next components and styles WIP
