+++ package.json
@@ -51,6 +51,7 @@
     "json-loader": "^0.5.4",
     "lodash": "^4.17.4",
     "lodash.take": "^4.1.1",
+    "marked": "^0.3.6",
     "md5": "^2.2.1",
     "mkdirp": "^0.5.1",
     "mocha": "^3.0.2",
