+++ site/components/date-bubble/style.css
@@ -1,7 +1,7 @@
-@value badgerRed from "../../../../../site/css/_colors.css";
+@value badgerRed from "../../css/_colors.css";
 
 .dateBubble {
-  composes: sansSerif from "../../../../../site/css/typography/_fonts.css";
+  composes: sansSerif from "../../css/typography/_fonts.css";
   color: badgerRed;
   text-transform: uppercase;
   min-width: 3em;
