+++ site/components/event-title/index.js
@@ -1,9 +1,9 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
 import icons from '../icons/style.css';
-import styles from '../events-list/style.css';
+import styles from '../../pages/events/events-list/style.css';
 import { h2 } from '../typography/style.css';
-import Link from '../../../../../site/components/link';
+import Link from '../link';
 
 const EventTitle = ({
   eventLink,
