+++ site/pages/events/events-news-list-entry/index.js
@@ -36,7 +36,7 @@ const EventsNewsListEntry = ({
       <Grid fit={false}>
         <Cell size={12}>
           <HR
-            color="grey" 
+            color="grey"
             customClassName={styles.mobileHorizontalLine}
           />
           <DateBubble
