+++ site/pages/events/style.css
@@ -1,4 +1,4 @@
-@value badgerWhite from "../../../../../site/css/_colors.css";
+@value badgerWhite from "../../css/_colors.css";
 
 .h1 {
   composes: h1 from "../../components/typography/style.css";
