+++ site/pages/job/style.css
@@ -1,5 +1,5 @@
 @value tablet from "../../components/variables/breakpoints.css";
-@value badgerWhite, badgerBlack, badgerRed from "../../../../../site/css/_colors.css";
+@value badgerWhite, badgerBlack, badgerRed from "../../css/_colors.css";
 
 .background {
   background-color: badgerWhite;
@@ -11,7 +11,7 @@
 }
 
 .applyLink {
-  composes: sansSerif from "../../../../../site/css/typography/_fonts.css";
+  composes: sansSerif from "../../css/typography/_fonts.css";
 
   float: right;
   display: none;
@@ -22,15 +22,15 @@
 }
 
 .sidebarApplyLink {
-  composes: smallButton from "../../../../../site/css/_links.css";
+  composes: smallButton from "../../css/_links.css";
 }
 
 .linkBack {
-  composes: bodyLink from "../../../../../site/css/_links.css";
+  composes: bodyLink from "../../css/_links.css";
   border-bottom: 1px solid badgerRed;
   display: inline-block;
   margin-bottom: 25px;
-  composes: sansSerif from "../../../../../site/css/typography/_fonts.css";
+  composes: sansSerif from "../../css/typography/_fonts.css";
 
   @media tablet {
     display: inline;
